requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.7
requests==2.20.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
